rename work load attestor modal
